``` Refactor service layer for resource and component management
- Introduced `component_notification.go` for notification logic.
- Added `resource_heartbeat.go` for heartbeat-specific operations.
- Added `resource_lifecycle.go` for resource monitoring lifecycle.
- Added `resource_validation.go` for validation logic.
- Moved status page analytics to `statuspage_analytics.go`.
- Updated `feature.json` to reflect new service decomposition.
- Updated `CLAUDE.md` and `roadmap.md` for alignment with changes.
  ```
