+++ .specify/feature.json
@@ -1,3 +1,3 @@
 {
-  "feature_directory": "specs/036-port-extraction"
+  "feature_directory": "specs/037-service-decomposition"
 }
